+++ app/javascript/components/game/map/MapDeployDisplay.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import MapHexPatterns from "./MapHexPatterns";
-import { Coordinate, Player } from "../../../utilities/commonTypes";
+import { Coordinate, ExtendedDirection, Player } from "../../../utilities/commonTypes";
 import Scenario from "../../../engine/Scenario";
 import { deployHex } from "../../../engine/control/deploy";
 import { DeployHexesTurn } from "../../../engine/Map";
@@ -14,6 +14,7 @@ interface MapDeployDisplayProps {
 
 export default function MapDeployDisplay({ scenario, scale, turn, player }: MapDeployDisplayProps) {
   const [hexDisplay, setHexDisplay] = useState<JSX.Element[]>([])
+  const [isolatedHexDisplay, setIsolatedHexDisplay] = useState<JSX.Element[]>([])
   const [vpDisplay, setVPDisplay] = useState<JSX.Element[]>([])
 
   const width = scenario.map.xSize * scale
@@ -21,9 +22,11 @@ export default function MapDeployDisplay({ scenario, scale, turn, player }: MapD
 
   useEffect(() => {
     const hexLoader: JSX.Element[] = []
+    const isolatedHexLoader: JSX.Element[] = []
     const vpLoader: JSX.Element[] = []
     scenario.map.mapHexes.forEach((row, y) => {
       row.forEach((hex, x) => {
+        const loc = new Coordinate(x, y)
         const hexes = player === 1 ? (scenario.map.alliedSetupHexes as DeployHexesTurn)[turn] :
           (scenario.map.axisSetupHexes as DeployHexesTurn)[turn]
         const show = deployHex(hexes, x, y) ? scenario.map.baseTerrainColor : "#AAA"
@@ -32,7 +35,23 @@ export default function MapDeployDisplay({ scenario, scale, turn, player }: MapD
           <polygon key={`${x}-${y}-h`} points={hex.hexCoords}
                    style={{ fill: show, stroke: showStroke, strokeWidth: 2 }} />
         )
-        const vp = scenario.map.victoryAt(new Coordinate(x, y)) 
+        const neighbors = scenario.map.hexNeighbors(loc)
+        if (deployHex(hexes, x, y)) {
+          let isolated = true
+          for (const n of neighbors) {
+            if (n && deployHex(hexes, n.coord.x, n.coord.y)) { isolated = false }
+          }
+          if (isolated) {
+            const coords = [0, 1, 2, 3, 4, 5, 6].map(i => {
+              return `${hex.xCorner(i as ExtendedDirection, -24)},${hex.yCorner(i as ExtendedDirection, -24)}`
+            }).join(" ")
+            isolatedHexLoader.push(
+              <polygon key={`${x}-${y}-h`} points={coords}
+                      style={{ fill: scenario.map.baseTerrainColor, strokeWidth: 0 }} />
+            )
+          }
+        }
+        const vp = scenario.map.victoryAt(loc)
         if (vp) {
           const fill = `url(#nation-${vp === 1 ? scenario.alliedFactions[0] : scenario.axisFactions[0]}-16)`
           vpLoader.push(
@@ -43,6 +62,7 @@ export default function MapDeployDisplay({ scenario, scale, turn, player }: MapD
       })
     })
     setHexDisplay(hexLoader)
+    setIsolatedHexDisplay(isolatedHexLoader)
     setVPDisplay(vpLoader)
   }, [scenario])
 
@@ -51,6 +71,7 @@ export default function MapDeployDisplay({ scenario, scale, turn, player }: MapD
          viewBox={`0 0 ${width / scale} ${height / scale}`}>
       <MapHexPatterns map={scenario.map} />
       {hexDisplay}
+      {isolatedHexDisplay}
       {vpDisplay}
     </svg>
   )